+++ test/test.js
@@ -13,9 +13,10 @@ describe('background-position', function () {
   it('should flip direction', function () {
     ensure('p{background-position:left top;}', 'p{background-position:right top;}');
     ensure('p{background-position:20px;}', 'p{background-position:right 20px;}');
-    ensure('p{background-position:20% top;}', 'p{background-position:80% top;}');
-    ensure('p{background-position:0 0;}', 'p{background-position:100% 0;}');
+    ensure('p{background-position:0% 0;}', 'p{background-position:100% 0;}');
     ensure('p{background-position:24.5% 0;}', 'p{background-position:75.5% 0;}');
+    ensure('p{background-position:20% top;}', 'p{background-position:80% top;}');
+    ensure('p{background-position:100% 0;}', 'p{background-position:0% 0;}');
 
     ensure('p{background-position-x:20%;}', 'p{background-position-x:80%;}');
     ensure('p{-ms-background-position-x:30%;}', 'p{-ms-background-position-x:70%;}');